add new_and_autosave.py
I had GIMP crash on me for a file I hadn't saved yet, everything lost,
fun times!

Anyway, new_and_autosave.py is bound to M-n and creates a new file,
saves it without prompting, then autosaves it every 2 mins. NOTE: this
does not set the image as clean, so when you close you'll still be
prompted to save the file. IMO, that is a good thing.